+++ subscriptions/app/application/adapters/EventBusInterface.php
@@ -11,4 +11,18 @@ interface EventBusInterface
      * @return void
      */
     public function publish(EventInterface $event): void;
+
+    /**
+     * @param string $queue
+     * @param string $routing
+     * @return void
+     */
+    public function subscribe(string $queue, string $routing): void;
+
+    /**
+     * @param string $queue
+     * @param string $routing
+     * @return void
+     */
+    public function unsubscribe(string $queue, string $routing): void;
 }